Improve aarch64 heterogeneous kernel tests Move aarch64 u8x16 ZipUnzip
