+++ diskann-wide/src/lib.rs
@@ -195,6 +195,9 @@ pub const ARCH: arch::Current = arch::current();
 ///    ```text
 ///    use of generic parameter from outer item
 ///    ```
+/// 5. `diskann_wide::alias!(f32s = type a::b::f32x4) => type f32s<A> = a::b::f32x4`:
+///    Use a direct path to the type rather than reaching the type as an associated type of
+///    [`Architecture`].
 #[macro_export]
 macro_rules! alias {
     ($var:ident) => {
@@ -210,6 +213,10 @@ macro_rules! alias {
     ($var:ident<$arch:ident> = $type:ident) => {
         #[allow(non_camel_case_types)]
         type $var<$arch> = <$arch as $crate::Architecture>::$type;
+    };
+    ($var:ident = type $type:path) => {
+        #[allow(non_camel_case_types)]
+        type $var = $type;
     };
 }
